requirements.txt
feedparser==5.2.1
python-dotenv==0.10.1